+++ lib/src/errors/models/http_response_exception.dart
@@ -1,11 +1,18 @@
+import 'package:core/src/errors/data/client_errors_map.dart';
+import 'package:core/src/errors/data/server_errors_map.dart';
+
 /// Simple [Exception] implementation for HTTP Response Error.
 class HttpResponseException implements Exception {
   final int statusCode;
-  final String name;
+  late String name;
   final String? message;
 
   /// An HTTP Response Exception.
-  HttpResponseException(this.statusCode, this.name, {this.message});
+  HttpResponseException(this.statusCode, {this.message}) {
+    final errorResponsesMap = clientErrorsMap..addAll(serverErrorsMap);
+
+    name = errorResponsesMap[statusCode] ?? 'undefined responses';
+  }
 
   @override
   String toString() => 'HttpResponseException: $statusCode ($name)';